chore: extract render helpers and fix viewport height
